refactor: Move data vis components into their own directory
