+++ packages/tools/devtools/devtools-view/src/components/data-visualization/FluidHandleView.tsx
@@ -3,6 +3,7 @@
  * Licensed under the MIT License.
  */
 import React from "react";
+
 import {
 	DataVisualization,
 	GetDataVisualization,
@@ -13,10 +14,11 @@ import {
 	HasFluidObjectId,
 	FluidObjectNode,
 } from "@fluid-experimental/devtools-core";
-import { useMessageRelay } from "../MessageRelayContext";
-import { Waiting } from "./Waiting";
+
+import { HasLabel } from "../CommonInterfaces";
+import { useMessageRelay } from "../../MessageRelayContext";
+import { Waiting } from "../Waiting";
 import { TreeDataView } from "./TreeDataView";
-import { HasLabel } from "./CommonInterfaces";
 
 const loggingContext = "EXTENSION(HandleView)";
