+++ packages/tools/devtools/devtools-view/src/components/data-visualization/FluidValueView.tsx
@@ -3,23 +3,25 @@
  * Licensed under the MIT License.
  */
 import React from "react";
-import { ValueNodeBase } from "@fluid-experimental/devtools-core";
 // eslint-disable-next-line import/no-internal-modules
 import { TreeItem } from "@fluentui/react-components/unstable";
+
+import { FluidObjectValueNode } from "@fluid-experimental/devtools-core";
+
+import { HasLabel } from "../CommonInterfaces";
 import { TreeHeader } from "./TreeHeader";
-import { HasLabel } from "./CommonInterfaces";
 
 /**
  * {@link ValueView} input props.
  */
-export interface ValueViewProps extends HasLabel {
-	node: ValueNodeBase;
+export interface FluidValueViewProps extends HasLabel {
+	node: FluidObjectValueNode;
 }
 
 /**
- * Render data with type VisualNodeKind.ValueNode and render its children.
+ * Render data with type VisualNodeKind.FluidValueNode and render its children.
  */
-export function ValueView(props: ValueViewProps): React.ReactElement {
+export function FluidValueView(props: FluidValueViewProps): React.ReactElement {
 	const { label, node } = props;
 
 	return (
