+++ packages/tools/devtools/devtools-view/src/components/data-visualization/UnknownDataView.tsx
@@ -3,6 +3,7 @@
  * Licensed under the MIT License.
  */
 import React from "react";
+
 import { UnknownObjectNode } from "@fluid-experimental/devtools-core";
 
 /**
